Moviendo query al archivo encargado de manejar la DB
